requirements.txt
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.2
idna==3.10
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pytest==8.3.3
pytest-cov==5.0.0
requests==2.32.3
urllib3==2.2.3
psycopg2-binary==2.9.9
config==0.5.1

